+++ packages/server-utils/src/ai/index.ts
@@ -8,3 +8,5 @@ export { instrumentOpenAiClient } from './openai';
 export { instrumentAnthropicAiClient } from './anthropic-ai';
 export { instrumentGoogleGenAIClient } from './google-genai';
 export { instrumentWorkersAiClient } from './workers-ai';
+export { createLangChainCallbackHandler, instrumentLangChainEmbeddings } from './langchain';
+export { instrumentStateGraph, instrumentStateGraphCompile, instrumentCreateReactAgent } from './langgraph';
