+++ packages/vercel-edge/src/index.ts
@@ -73,7 +73,6 @@ export {
   functionToStringIntegration,
   // eslint-disable-next-line typescript/no-deprecated
   inboundFiltersIntegration,
-  instrumentStateGraph,
   eventFiltersIntegration,
   linkedErrorsIntegration,
   requestDataIntegration,
@@ -97,8 +96,6 @@ export {
   wrapMcpServerWithSentry,
   consoleLoggingIntegration,
   createConsolaReporter,
-  createLangChainCallbackHandler,
-  instrumentLangChainEmbeddings,
   featureFlagsIntegration,
   logger,
   metrics,
@@ -109,6 +106,9 @@ export {
   instrumentOpenAiClient,
   instrumentAnthropicAiClient,
   instrumentGoogleGenAIClient,
+  createLangChainCallbackHandler,
+  instrumentLangChainEmbeddings,
+  instrumentStateGraph,
 } from '@sentry/server-utils/no-diagnostic-channels';
 
 export { VercelEdgeClient } from './client';
